// src/app/shared/enums/app-api-urls.enum.ts
export enum AppApiUrls {
  // Controls all api of application

  PARENT = ".",
  URL_SEPARATER = "/",

  // API Common URLs
  SIGN_IN_URL = "/api/signin",
  CHANGE_PASS_URL = "/api/changePassword",
  PATIENT_USER_URL = "/api/user/patient",
  PATIENT_URL = "/api/patient",
  ADMIN_URL = "/api/admin",
  DOCTOR_URL = "/api/doctor",
  CM_URL = "/api/caremanager",
  SERVICE_CATEGORY_URL = "/api/category",
  QUESTIONAIR_URL = "/api/questionnaire",
  CASE_URL = "/api/case",
  PATIENT_SIGN_UP_URL = "/api/patient/signup",
  EARNING_SERVICE_URL = "/api/payments"
}